Main Project
Project: Comparative Study between Linear Model and Non Linear Models along with Ensemble Models on Regression Datasets
Description: The motivation of the project is, compare the linear model and the non-linear models along with ensemble models on regression datasets. With the help of R studio, comparing the six machine learning techniques efficiency with different datasets. The statistical models under consideration are; Multiple linear regression is under the linear model, Decision tree ,support vector machine and k-nearest neighbours are the non-linear models and Random forest and the extreme gradient boosting are the ensemble models.